Final Deliverables/main/requirements.txt
flask
ibm_db
ibm-cos-sdk
pyjwt
requests
trycourier